requirements.txt
discord.py
discord-py-slash-command
python-dotenv~=0.15.0
firebase_admin
disputils~=0.1.3
urllib3~=1.26.3
requests~=2.25.1
discord~=1.0.1
gspread~=4.0.1
oauth2client~=4.1.3
google-api-python-client~=2.21.0
youtube-search~=2.1.0
